Site Editor: Remove useEditedEntityRecord hook
